// src/types/types.ts
export type GeoCoordinates = {
  latitude: number; // Latitude
  longitude: number; // Longitude
};
// ...
export type CityData = {
  name: string;
  latitude: number;
  longitude: number;
  country: string;
  state?: string;
  id: string;
};